+++ STACpopulator/extensions/rdps.py
@@ -1,0 +1,13 @@
+from __future__ import annotations
+
+from STACpopulator.extensions.cf import CFHelper
+from STACpopulator.extensions.file import FileHelper
+from STACpopulator.extensions.thredds import THREDDSCatalogDataModel
+
+
+class RDPSDataModel(THREDDSCatalogDataModel):
+    """Data model for RDPS NetCDF datasets."""
+
+    # Extension classes
+    cf: CFHelper
+    file: FileHelper