+++ CHANGES.md
@@ -4,8 +4,11 @@
 
 <!-- insert list items of new changes here -->
 
-## [0.10.0](https://github.com/crim-ca/stac-populator/tree/0.10.0) (2025-11-11)
+## [1.11.0](https://github.com/crim-ca/stac-populator/tree/1.11.0) (2025-11-12)
 
+* Add option to automatically update collection extents and summaries based on ingested items.
+
+## [0.10.0](https://github.com/crim-ca/stac-populator/tree/0.10.0) (2025-11-11)
 
 <!-- insert list items of new changes here -->